+++ src/components/react_canvas/utils/rendering_props_utils.test.ts
@@ -5,19 +5,24 @@ import {
   buildBarProps,
   buildLinePointProps,
   buildLineProps,
+  buildPointStyleProps,
 } from './rendering_props_utils';
 
 describe('[canvas] Area Geometries props', () => {
   test('can build area point props', () => {
+    const pointStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 2,
+      opacity: 0.5,
+    });
+
     const props = buildAreaPointProps({
       areaIndex: 1,
       pointIndex: 2,
       x: 10,
       y: 20,
-      radius: 30,
-      strokeWidth: 2,
       color: 'red',
-      opacity: 0.5,
+      pointStyleProps,
     });
     expect(props).toEqual({
       key: 'area-point-1-2',
@@ -34,15 +39,19 @@ describe('[canvas] Area Geometries props', () => {
       listening: false,
     });
 
+    const noStrokePointStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 0,
+      opacity: 0.5,
+    });
+
     const propsNoStroke = buildAreaPointProps({
       areaIndex: 1,
       pointIndex: 2,
       x: 10,
       y: 20,
-      radius: 30,
-      strokeWidth: 0,
       color: 'red',
-      opacity: 0.5,
+      pointStyleProps: noStrokePointStyleProps,
     });
     expect(propsNoStroke).toEqual({
       key: 'area-point-1-2',
@@ -58,21 +67,83 @@ describe('[canvas] Area Geometries props', () => {
       perfectDrawEnabled: false,
       listening: false,
     });
+
+    const seriesPointStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 2,
+      opacity: 0.5,
+      seriesPointStyle: {
+        radius: 123,
+        stroke: 'series-stroke',
+        strokeWidth: 456,
+        opacity: 789,
+        visible: true,
+      },
+    });
+    const seriesPointStyle = buildAreaPointProps({
+      areaIndex: 1,
+      pointIndex: 2,
+      x: 10,
+      y: 20,
+      color: 'red',
+      pointStyleProps: seriesPointStyleProps,
+    });
+    expect(seriesPointStyle).toEqual({
+      key: 'area-point-1-2',
+      x: 10,
+      y: 20,
+      radius: 123,
+      strokeWidth: 456,
+      strokeEnabled: true,
+      stroke: 'red',
+      fill: 'white',
+      opacity: 789,
+      strokeHitEnabled: false,
+      perfectDrawEnabled: false,
+      listening: false,
+    });
   });
   test('can build area path props', () => {
     const props = buildAreaProps({
       index: 1,
       areaPath: 'M0,0L10,10Z',
+      xTransform: 40,
       color: 'red',
       opacity: 0.5,
     });
     expect(props).toEqual({
       key: 'area-1',
       data: 'M0,0L10,10Z',
+      x: 40,
       fill: 'red',
       lineCap: 'round',
       lineJoin: 'round',
       opacity: 0.5,
+      strokeHitEnabled: false,
+      perfectDrawEnabled: false,
+      listening: false,
+    });
+
+    const seriesAreaStyle = buildAreaProps({
+      index: 1,
+      areaPath: 'M0,0L10,10Z',
+      xTransform: 0,
+      color: 'red',
+      opacity: 0.5,
+      seriesAreaStyle: {
+        opacity: 123,
+        fill: '',
+        visible: true,
+      },
+    });
+    expect(seriesAreaStyle).toEqual({
+      key: 'area-1',
+      data: 'M0,0L10,10Z',
+      x: 0,
+      fill: 'red',
+      lineCap: 'round',
+      lineJoin: 'round',
+      opacity: 123,
       strokeHitEnabled: false,
       perfectDrawEnabled: false,
       listening: false,
@@ -82,6 +153,7 @@ describe('[canvas] Area Geometries props', () => {
     const props = buildAreaLineProps({
       areaIndex: 1,
       lineIndex: 2,
+      xTransform: 40,
       linePath: 'M0,0L10,10Z',
       color: 'red',
       strokeWidth: 1,
@@ -92,6 +164,7 @@ describe('[canvas] Area Geometries props', () => {
     expect(props).toEqual({
       key: `area-1-line-2`,
       data: 'M0,0L10,10Z',
+      x: 40,
       stroke: 'red',
       strokeWidth: 1,
       lineCap: 'round',
@@ -102,20 +175,55 @@ describe('[canvas] Area Geometries props', () => {
       listening: false,
     });
     expect(props.fill).toBeFalsy();
+
+    const seriesLineStyle = buildAreaLineProps({
+      areaIndex: 1,
+      xTransform: 0,
+      lineIndex: 2,
+      linePath: 'M0,0L10,10Z',
+      color: 'red',
+      strokeWidth: 1,
+      geometryStyle: {
+        opacity: 0.5,
+      },
+      seriesAreaLineStyle: {
+        opacity: 0.5,
+        stroke: 'series-stroke',
+        strokeWidth: 66,
+        visible: true,
+      },
+    });
+    expect(seriesLineStyle).toEqual({
+      key: `area-1-line-2`,
+      data: 'M0,0L10,10Z',
+      x: 0,
+      stroke: 'red',
+      strokeWidth: 66,
+      lineCap: 'round',
+      lineJoin: 'round',
+      opacity: 0.5,
+      strokeHitEnabled: false,
+      perfectDrawEnabled: false,
+      listening: false,
+    });
   });
 });
 
 describe('[canvas] Line Geometries', () => {
   test('can build line point props', () => {
+    const pointStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 2,
+      opacity: 0.5,
+    });
+
     const props = buildLinePointProps({
       lineIndex: 1,
       pointIndex: 2,
       x: 10,
       y: 20,
-      radius: 30,
-      strokeWidth: 2,
       color: 'red',
-      opacity: 0.5,
+      pointStyleProps,
     });
     expect(props).toEqual({
       key: 'line-point-1-2',
@@ -132,15 +240,18 @@ describe('[canvas] Line Geometries', () => {
       listening: false,
     });
 
+    const noStrokeStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 0,
+      opacity: 0.5,
+    });
     const propsNoStroke = buildLinePointProps({
       lineIndex: 1,
       pointIndex: 2,
       x: 10,
       y: 20,
-      radius: 30,
-      strokeWidth: 0,
       color: 'red',
-      opacity: 0.5,
+      pointStyleProps: noStrokeStyleProps,
     });
     expect(propsNoStroke).toEqual({
       key: 'line-point-1-2',
@@ -156,14 +267,49 @@ describe('[canvas] Line Geometries', () => {
       perfectDrawEnabled: false,
       listening: false,
     });
+
+    const seriesPointStyleProps = buildPointStyleProps({
+      radius: 30,
+      strokeWidth: 2,
+      opacity: 0.5,
+      seriesPointStyle: {
+        stroke: 'series-stroke',
+        strokeWidth: 6,
+        visible: true,
+        radius: 12,
+        opacity: 18,
+      },
+    });
+    const seriesPointStyle = buildLinePointProps({
+      lineIndex: 1,
+      pointIndex: 2,
+      x: 10,
+      y: 20,
+      color: 'red',
+      pointStyleProps: seriesPointStyleProps,
+    });
+    expect(seriesPointStyle).toEqual({
+      key: 'line-point-1-2',
+      x: 10,
+      y: 20,
+      radius: 12,
+      strokeWidth: 6,
+      strokeEnabled: true,
+      stroke: 'red',
+      fill: 'white',
+      opacity: 18,
+      strokeHitEnabled: false,
+      perfectDrawEnabled: false,
+      listening: false,
+    });
   });
   test('can build line path props', () => {
     const props = buildLineProps({
       index: 1,
       linePath: 'M0,0L10,10Z',
+      xTransform: 40,
       color: 'red',
       strokeWidth: 1,
-      opacity: 0.3,
       geometryStyle: {
         opacity: 0.5,
       },
@@ -171,6 +317,7 @@ describe('[canvas] Line Geometries', () => {
     expect(props).toEqual({
       key: `line-1`,
       data: 'M0,0L10,10Z',
+      x: 40,
       stroke: 'red',
       strokeWidth: 1,
       lineCap: 'round',
@@ -181,6 +328,35 @@ describe('[canvas] Line Geometries', () => {
       listening: false,
     });
     expect(props.fill).toBeFalsy();
+
+    const seriesLineStyleProps = buildLineProps({
+      index: 1,
+      linePath: 'M0,0L10,10Z',
+      xTransform: 0,
+      color: 'red',
+      strokeWidth: 1,
+      geometryStyle: {
+        opacity: 0.5,
+      },
+      seriesLineStyle: {
+        stroke: 'series-stroke',
+        strokeWidth: 66,
+        visible: true,
+      },
+    });
+    expect(seriesLineStyleProps).toEqual({
+      key: `line-1`,
+      data: 'M0,0L10,10Z',
+      x: 0,
+      stroke: 'red',
+      strokeWidth: 66,
+      lineCap: 'round',
+      lineJoin: 'round',
+      opacity: 0.5,
+      strokeHitEnabled: false,
+      perfectDrawEnabled: false,
+      listening: false,
+    });
   });
 });