+++ src/lib/series/specs.ts
@@ -1,4 +1,10 @@
-import { AnnotationLineStyle, GridLineConfig } from '../themes/theme';
+import {
+  AnnotationLineStyle,
+  AreaSeriesStyle,
+  CustomBarSeriesStyle,
+  GridLineConfig,
+  LineSeriesStyle,
+} from '../themes/theme';
 import { Accessor } from '../utils/accessor';
 import { AnnotationId, AxisId, GroupId, SpecId } from '../utils/ids';
 import { ScaleContinuousType, ScaleType } from '../utils/scales/scales';
@@ -36,7 +42,7 @@ export interface SeriesSpec {
   /** An array of data */
   data: Datum[];
   /** The type of series you are looking to render */
-  seriesType: 'bar' | 'line' | 'area' | 'basic';
+  seriesType: 'bar' | 'line' | 'area';
   /** Custom colors for series */
   customSeriesColors?: CustomSeriesColorsMap;
   /** If the series should appear in the legend
@@ -84,11 +90,18 @@ export interface SeriesScales {
    * @default ScaleType.Linear
    */
   yScaleType: ScaleContinuousType;
-  /** if true, the min y value is set to the minimum domain value, 0 otherwise */
+  /**
+   * if true, the min y value is set to the minimum domain value, 0 otherwise
+   * @default false
+   */
   yScaleToDataExtent: boolean;
 }
 
-export type BasicSeriesSpec = SeriesSpec & SeriesAccessors & SeriesScales;
+export type BasicSeriesSpec = SeriesSpec & SeriesAccessors & SeriesScales & {
+  barSeriesStyle?: CustomBarSeriesStyle;
+  lineSeriesStyle?: LineSeriesStyle;
+  areaSeriesStyle?: AreaSeriesStyle;
+};
 
 /**
  * This spec describe the dataset configuration used to display a bar series.