+++ src/state/utils.ts
@@ -11,7 +11,6 @@ import {
   renderArea,
   renderBars,
   renderLine,
-  renderPoints,
 } from '../lib/series/rendering';
 import { computeXScale, computeYScales, countBarsInCluster } from '../lib/series/scales';
 import {
@@ -309,7 +308,7 @@ export function renderGeometries(
   const bars: BarGeometry[] = [];
   const areas: AreaGeometry[] = [];
   const lines: LineGeometry[] = [];
-  let pointGeometriesIndex: Map<any, IndexedGeometry[]> = new Map();
+  const pointGeometriesIndex: Map<any, IndexedGeometry[]> = new Map();
   let barGeometriesIndex: Map<any, IndexedGeometry[]> = new Map();
   let lineGeometriesIndex: Map<any, IndexedGeometry[]> = new Map();
   let areaGeometriesIndex: Map<any, IndexedGeometry[]> = new Map();
@@ -329,29 +328,13 @@ export function renderGeometries(
     }
     const color = seriesColorsMap.get(ds.seriesColorKey) || defaultColor;
     switch (spec.seriesType) {
-      case 'basic':
-        const pointShift = clusteredCount > 0 ? clusteredCount : 1;
-
-        const renderedPoints = renderPoints(
-          (xScale.bandwidth * pointShift) / 2,
-          ds.data,
-          xScale,
-          yScale,
-          color,
-          ds.specId,
-          Boolean(spec.y0Accessors),
-          ds.key,
-        );
-        points.push(...renderedPoints.pointGeometries);
-        pointGeometriesIndex = mergeGeometriesIndexes(
-          pointGeometriesIndex,
-          renderedPoints.indexedGeometries,
-        );
-        geometriesCounts.points += renderedPoints.pointGeometries.length;
-        break;
       case 'bar':
         const shift = isStacked ? indexOffset : indexOffset + i;
-        const renderedBars = renderBars(shift, ds.data, xScale, yScale, color, ds.specId, ds.key, spec.showValueLabel);
+        const barSeriesStyle = spec.barSeriesStyle;
+        const renderedBars = renderBars(
+          shift, ds.data, xScale, yScale, color,
+          ds.specId, ds.key, spec.showValueLabel, barSeriesStyle,
+        );
         barGeometriesIndex = mergeGeometriesIndexes(
           barGeometriesIndex,
           renderedBars.indexedGeometries,
@@ -361,6 +344,7 @@ export function renderGeometries(
         break;
       case 'line':
         const lineShift = clusteredCount > 0 ? clusteredCount : 1;
+        const lineSeriesStyle = spec.lineSeriesStyle;
         const renderedLines = renderLine(
           // move the point on half of the bandwidth if we have mixed bars/lines
           (xScale.bandwidth * lineShift) / 2,
@@ -372,6 +356,7 @@ export function renderGeometries(
           ds.specId,
           Boolean(spec.y0Accessors),
           ds.key,
+          lineSeriesStyle,
         );
         lineGeometriesIndex = mergeGeometriesIndexes(
           lineGeometriesIndex,
@@ -383,6 +368,7 @@ export function renderGeometries(
         break;
       case 'area':
         const areaShift = clusteredCount > 0 ? clusteredCount : 1;
+        const areaSeriesStyle = spec.areaSeriesStyle;
         const renderedAreas = renderArea(
           // move the point on half of the bandwidth if we have mixed bars/lines
           (xScale.bandwidth * areaShift) / 2,
@@ -394,6 +380,7 @@ export function renderGeometries(
           ds.specId,
           Boolean(spec.y0Accessors),
           ds.key,
+          areaSeriesStyle,
         );
         areaGeometriesIndex = mergeGeometriesIndexes(
           areaGeometriesIndex,