+++ stories/bar_chart.tsx
@@ -26,17 +26,22 @@ storiesOf('Bar Chart', module)
   .add('basic', () => {
     const darkmode = boolean('darkmode', false);
     const className = darkmode ? 'story-chart-dark' : 'story-chart';
+    const toggleSpec = boolean('toggle bar spec', true);
+    const data1 = [{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }];
+    const data2 = data1.map((datum) => ({ ...datum, y: datum.y - 1 }));
+    const data = toggleSpec ? data1 : data2;
+    const specId = toggleSpec ? 'bars1' : 'bars2';
+
     return (
-      <Chart renderer="canvas" className={className}>
+      <Chart className={className}>
         <BarSeries
-          id={getSpecId('bars')}
+          id={getSpecId(specId)}
           name={'Simple bar series'}
           xScaleType={ScaleType.Linear}
           yScaleType={ScaleType.Linear}
           xAccessor="x"
           yAccessors={['y']}
-          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
+          data={data}
         />
       </Chart>
     );
@@ -76,7 +81,7 @@ storiesOf('Bar Chart', module)
     const className = darkmode ? 'story-chart-dark' : 'story-chart';
     const defaultTheme = darkmode ? DARK_THEME : LIGHT_THEME;
     return (
-      <Chart renderer="canvas" className={className}>
+      <Chart className={className}>
         <Settings theme={defaultTheme} />
         <Axis
           id={getAxisId('bottom')}
@@ -98,14 +103,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('with ordinal x axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -126,14 +130,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 'a', y: 2 }, { x: 'b', y: 7 }, { x: 'c', y: 3 }, { x: 'd', y: 6 }]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('with linear x axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -154,7 +157,6 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 1, y: 2 }, { x: 2, y: 7 }, { x: 4, y: 3 }, { x: 9, y: 6 }]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -162,7 +164,7 @@ storiesOf('Bar Chart', module)
   .add('with time x axis', () => {
     const formatter = timeFormatter(niceTimeFormatByDay(1));
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings debug={boolean('debug', false)} />
         <Axis
           id={getAxisId('bottom')}
@@ -186,14 +188,13 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('with log y axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -230,7 +231,7 @@ storiesOf('Bar Chart', module)
   })
   .add('with stacked log y axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -277,7 +278,7 @@ storiesOf('Bar Chart', module)
   })
   .add('with axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -300,14 +301,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('stacked with axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -340,14 +340,13 @@ storiesOf('Bar Chart', module)
             { x: 2, y: 8, g: 'b' },
             { x: 3, y: 2, g: 'b' },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('clustered with axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -379,14 +378,13 @@ storiesOf('Bar Chart', module)
             { x: 2, y: 8, g: 'b' },
             { x: 3, y: 2, g: 'b' },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('clustered multiple series specs', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -408,7 +406,6 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId('bar series 2')}
@@ -417,7 +414,6 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 0, y: 1 }, { x: 1, y: 2 }, { x: 2, y: 3 }, { x: 3, y: 4 }]}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId('bar series 3')}
@@ -436,7 +432,6 @@ storiesOf('Bar Chart', module)
             { x: 2, y: 9, g: 'b' },
             { x: 3, y: 2, g: 'b' },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -444,7 +439,7 @@ storiesOf('Bar Chart', module)
   .add('time clustered using various specs', () => {
     const formatter = timeFormatter(niceTimeFormatByDay(1));
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings debug={boolean('debug', false)} />
         <Axis
           id={getAxisId('bottom')}
@@ -468,7 +463,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId(KIBANA_METRICS.metrics.kibana_os_load[1].metric.label)}
@@ -477,7 +471,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[1].data}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId(KIBANA_METRICS.metrics.kibana_os_load[2].metric.label)}
@@ -486,7 +479,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[2].data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -494,7 +486,7 @@ storiesOf('Bar Chart', module)
   .add('time stacked using various specs', () => {
     const formatter = timeFormatter(niceTimeFormatByDay(1));
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings debug={boolean('debug', false)} />
         <Axis
           id={getAxisId('bottom')}
@@ -519,7 +511,6 @@ storiesOf('Bar Chart', module)
           yAccessors={[1]}
           stackAccessors={[0]}
           data={KIBANA_METRICS.metrics.kibana_os_load[2].data.slice(0, 20)}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId(KIBANA_METRICS.metrics.kibana_os_load[1].metric.label)}
@@ -529,7 +520,6 @@ storiesOf('Bar Chart', module)
           yAccessors={[1]}
           stackAccessors={[0]}
           data={KIBANA_METRICS.metrics.kibana_os_load[1].data.slice(0, 20)}
-          yScaleToDataExtent={false}
         />
         <BarSeries
           id={getSpecId(KIBANA_METRICS.metrics.kibana_os_load[0].metric.label)}
@@ -539,14 +529,13 @@ storiesOf('Bar Chart', module)
           yAccessors={[1]}
           stackAccessors={[0]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data.slice(0, 20)}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('1y0g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -568,14 +557,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={TestDatasets.BARCHART_1Y0G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('1y1g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -598,14 +586,13 @@ storiesOf('Bar Chart', module)
           yAccessors={['y']}
           splitSeriesAccessors={['g']}
           data={TestDatasets.BARCHART_1Y1G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('1y2g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -628,14 +615,13 @@ storiesOf('Bar Chart', module)
           yAccessors={['y']}
           splitSeriesAccessors={['g1', 'g2']}
           data={TestDatasets.BARCHART_1Y2G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('2y0g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -657,14 +643,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y1', 'y2']}
           data={TestDatasets.BARCHART_2Y0G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('2y1g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -687,14 +672,13 @@ storiesOf('Bar Chart', module)
           yAccessors={['y1', 'y2']}
           splitSeriesAccessors={['g']}
           data={TestDatasets.BARCHART_2Y1G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('2y2g', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -717,7 +701,6 @@ storiesOf('Bar Chart', module)
           yAccessors={['y1', 'y2']}
           splitSeriesAccessors={['g1', 'g2']}
           data={TestDatasets.BARCHART_2Y2G}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -726,7 +709,7 @@ storiesOf('Bar Chart', module)
     const dg = new DataGenerator();
     const data = dg.generateSimpleSeries(200);
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -742,14 +725,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('single data chart', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -765,14 +747,13 @@ storiesOf('Bar Chart', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 1, y: 10 }]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('single data clusterd chart', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -794,14 +775,13 @@ storiesOf('Bar Chart', module)
             { x: 1, y: 3, g: 'c' },
             { x: 1, y: 10, g: 'd' },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('single data stacked chart', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -824,14 +804,13 @@ storiesOf('Bar Chart', module)
             { x: 1, y: 3, g: 'c' },
             { x: 1, y: 10, g: 'd' },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('single data stacked chart scale to extent', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -856,7 +835,7 @@ storiesOf('Bar Chart', module)
   })
   .add('single data clustered chart scale to extent', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -880,7 +859,7 @@ storiesOf('Bar Chart', module)
   })
   .add('negative and positive x values', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -906,7 +885,6 @@ storiesOf('Bar Chart', module)
             { x: -1, y: 3 },
             { x: 3, y: 1 },
           ]}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -947,7 +925,7 @@ storiesOf('Bar Chart', module)
     }
 
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('top')} position={Position.Top} title={'Top axis'} />
         <Axis
           id={getAxisId('left2')}
@@ -983,7 +961,7 @@ storiesOf('Bar Chart', module)
     });
     const scaleToDataExtent = boolean('scale to extent', true);
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           title={'timestamp per 1 minute'}
@@ -1026,7 +1004,7 @@ storiesOf('Bar Chart', module)
   .add('[test] - linear', () => {
     const data = [[1, 1], [2, 2], [3, 3], [4, 4], [5, 5], [6, 4], [7, 3], [8, 2], [9, 1]];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} title={'index'} position={Position.Bottom} />
         <Axis
           id={getAxisId('left')}
@@ -1041,7 +1019,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -1060,7 +1037,7 @@ storiesOf('Bar Chart', module)
       [start.plus({ minute: 8 }).toMillis(), 1],
     ];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           title={'index'}
@@ -1080,7 +1057,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -1098,7 +1074,7 @@ storiesOf('Bar Chart', module)
       [9, 1, 3],
     ];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} title={'index'} position={Position.Bottom} />
         <Axis
           id={getAxisId('left')}
@@ -1113,7 +1089,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1, 2]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -1132,7 +1107,7 @@ storiesOf('Bar Chart', module)
       [start.plus({ minute: 8 }).toMillis(), 1, 4],
     ];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           title={'index'}
@@ -1152,7 +1127,6 @@ storiesOf('Bar Chart', module)
           xAccessor={0}
           yAccessors={[1, 2]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -1169,7 +1143,7 @@ storiesOf('Bar Chart', module)
       [4, 4, 6, 'b'],
     ];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} title={'index'} position={Position.Bottom} />
         <Axis
           id={getAxisId('left')}
@@ -1185,7 +1159,6 @@ storiesOf('Bar Chart', module)
           yAccessors={[1]}
           splitSeriesAccessors={[3]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -1202,7 +1175,7 @@ storiesOf('Bar Chart', module)
       [4, 4, 6, 'b'],
     ];
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis id={getAxisId('bottom')} title={'index'} position={Position.Bottom} />
         <Axis
           id={getAxisId('left')}
@@ -1219,7 +1192,6 @@ storiesOf('Bar Chart', module)
           splitSeriesAccessors={[3]}
           stackAccessors={[0]}
           data={data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );