+++ stories/line_chart.tsx
@@ -1,3 +1,4 @@
+import { boolean } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
 import {
@@ -15,27 +16,33 @@ import {
 } from '../src/';
 import { KIBANA_METRICS } from '../src/lib/series/utils/test_dataset_kibana';
 import { TSVB_DATASET } from '../src/lib/series/utils/test_dataset_tsvb';
+
 const dateFormatter = timeFormatter(niceTimeFormatByDay(1));
 
 storiesOf('Line Chart', module)
   .add('basic', () => {
+    const toggleSpec = boolean('toggle line spec', true);
+    const data1 = KIBANA_METRICS.metrics.kibana_os_load[0].data;
+    const data2 = data1.map((datum) => [datum[0], datum[1] - 1]);
+    const data = toggleSpec ? data1 : data2;
+    const specId = toggleSpec ? 'lines1' : 'lines2';
+
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <LineSeries
-          id={getSpecId('lines')}
+          id={getSpecId(specId)}
           xScaleType={ScaleType.Time}
           yScaleType={ScaleType.Linear}
           xAccessor={0}
           yAccessors={[1]}
-          data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
-          yScaleToDataExtent={false}
+          data={data}
         />
       </Chart>
     );
   })
   .add('w axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -55,14 +62,13 @@ storiesOf('Line Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data.slice(0, 5)}
-          yScaleToDataExtent={true}
         />
       </Chart>
     );
   })
   .add('ordinal w axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -82,14 +88,13 @@ storiesOf('Line Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data.slice(0, 5)}
-          yScaleToDataExtent={true}
         />
       </Chart>
     );
   })
   .add('linear w axis', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Axis
           id={getAxisId('bottom')}
           position={Position.Bottom}
@@ -109,14 +114,13 @@ storiesOf('Line Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data.slice(0, 5)}
-          yScaleToDataExtent={true}
         />
       </Chart>
     );
   })
   .add('w axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -137,14 +141,13 @@ storiesOf('Line Chart', module)
           xAccessor={0}
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('curved w axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -167,7 +170,6 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.CURVE_MONOTONE_X}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('basis')}
@@ -177,7 +179,6 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.CURVE_BASIS}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('cardinal')}
@@ -187,7 +188,6 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.CURVE_CARDINAL}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('catmull rom')}
@@ -197,7 +197,6 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.CURVE_CATMULL_ROM}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('natural')}
@@ -207,7 +206,6 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.CURVE_NATURAL}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('linear')}
@@ -217,14 +215,13 @@ storiesOf('Line Chart', module)
           yAccessors={[1]}
           data={KIBANA_METRICS.metrics.kibana_os_load[0].data}
           curve={CurveType.LINEAR}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
   })
   .add('multiple w axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -271,7 +268,7 @@ storiesOf('Line Chart', module)
   })
   .add('stacked w axis and legend', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -320,7 +317,7 @@ storiesOf('Line Chart', module)
   })
   .add('multi series with log values (limit 0 or negative values)', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -345,7 +342,6 @@ storiesOf('Line Chart', module)
               yAccessors={[1]}
               data={series.data}
               curve={CurveType.CURVE_MONOTONE_X}
-              yScaleToDataExtent={false}
             />
           );
         })}