+++ stories/mixed.tsx
@@ -6,6 +6,7 @@ import {
   Axis,
   BarSeries,
   Chart,
+  CurveType,
   getAxisId,
   getSpecId,
   LineSeries,
@@ -18,7 +19,7 @@ import { timeFormatter } from '../src/utils/data/formatters';
 storiesOf('Mixed Charts', module)
   .add('bar and lines', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis id={getAxisId('bottom')} position={Position.Bottom} title={'Bottom axis'} />
         <Axis
@@ -49,7 +50,7 @@ storiesOf('Mixed Charts', module)
   })
   .add('lines and areas', () => {
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -71,7 +72,6 @@ storiesOf('Mixed Charts', module)
           xAccessor="x"
           yAccessors={['y']}
           data={[{ x: 0, y: 2.5 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('lines')}
@@ -81,6 +81,60 @@ storiesOf('Mixed Charts', module)
           yAccessors={['y']}
           stackAccessors={['x']}
           data={[{ x: 0, y: 2.8 }, { x: 1, y: 2 }, { x: 2, y: 4 }, { x: 3, y: 10 }]}
+        />
+      </Chart>
+    );
+  })
+  .add('areas and bars', () => {
+    return (
+      <Chart renderer="canvas" className={'story-chart'}>
+        <Settings
+          showLegend={true}
+          legendPosition={Position.Right}
+        />
+        <Axis
+          id={getAxisId('bottom')}
+          position={Position.Bottom}
+          title={'Bottom axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('left2')}
+          title={'Left axis'}
+          position={Position.Left}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+        <Axis
+          id={getAxisId('top')}
+          position={Position.Top}
+          title={'Top axis'}
+          showOverlappingTicks={true}
+        />
+        <Axis
+          id={getAxisId('right')}
+          title={'Right axis'}
+          position={Position.Right}
+          tickFormat={(d) => Number(d).toFixed(2)}
+        />
+        <BarSeries
+          id={getSpecId('bars')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y']}
+          splitSeriesAccessors={['g']}
+          stackAccessors={['x']}
+          data={[{ x: 0, y: 2 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
+          yScaleToDataExtent={false}
+        />
+        <AreaSeries
+          id={getSpecId('areas')}
+          xScaleType={ScaleType.Linear}
+          yScaleType={ScaleType.Linear}
+          xAccessor="x"
+          yAccessors={['y']}
+          curve={CurveType.CURVE_MONOTONE_X}
+          data={[{ x: 0, y: 2.5 }, { x: 1, y: 7 }, { x: 2, y: 3 }, { x: 3, y: 6 }]}
           yScaleToDataExtent={false}
         />
       </Chart>
@@ -91,7 +145,7 @@ storiesOf('Mixed Charts', module)
     const data2 = [[1, 5], [2, 4], [3, 3], [4, 2], [5, 1], [6, 2], [7, 3], [8, 4], [9, 5]];
 
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -113,7 +167,6 @@ storiesOf('Mixed Charts', module)
           xAccessor={0}
           yAccessors={[1]}
           data={data1}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('data2')}
@@ -122,7 +175,6 @@ storiesOf('Mixed Charts', module)
           xAccessor={0}
           yAccessors={[1]}
           data={data2}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );
@@ -154,7 +206,7 @@ storiesOf('Mixed Charts', module)
     const dateFormatter = timeFormatter('HH:mm:ss');
 
     return (
-      <Chart renderer="canvas" className={'story-chart'}>
+      <Chart className={'story-chart'}>
         <Settings showLegend={true} legendPosition={Position.Right} />
         <Axis
           id={getAxisId('bottom')}
@@ -177,7 +229,6 @@ storiesOf('Mixed Charts', module)
           xAccessor={0}
           yAccessors={[1]}
           data={data1}
-          yScaleToDataExtent={false}
         />
         <LineSeries
           id={getSpecId('data2')}
@@ -186,7 +237,6 @@ storiesOf('Mixed Charts', module)
           xAccessor={0}
           yAccessors={[2]}
           data={data2}
-          yScaleToDataExtent={false}
         />
       </Chart>
     );